final population doubling calculated each count

+++ src/main/java/com/biotools/dto/ConditionDTO.java
@@ -1,12 +1,14 @@
 package com.biotools.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
+
 import lombok.Data;
 
 @Data
 public class ConditionDTO {
 	private String cellLine;
-	private int initialPopulationDoubling;
+	private BigDecimal initialPopulationDoubling;
 	private List<CellCountDTO> cellCountList;
 	private List<TreatmentDTO> treatmentList;
 }
